+++ .copr/local.sh
@@ -24,7 +24,7 @@ cd "$repodir" || exit
 # We could download the correct git tag as a tar.gz file, but we already have all the files here! Why do that?
 # Make a directory, copy the needed files of the repo into the directory, then tar the results.
 mkdir -p "$workingdir/XIVLauncher4rpm-$DownstreamTag"
-cp cleanupprofile.sh openssl_fix.cnf xivlauncher.sh XIVLauncher.desktop COPYING "$workingdir/XIVLauncher4rpm-$DownstreamTag/"
+cp cleanupprofile.sh openssl_fix.cnf xivlauncher.sh XIVLauncher.desktop XIVLauncher-custom.desktop COPYING "$workingdir/XIVLauncher4rpm-$DownstreamTag/"
 cd "$workingdir" || exit
 echo "Writing XIVLauncher4rpm-$DownstreamTag.tar.xz"
 tar --totals -I 'xz -v -9' -cf "$xlsource/XIVLauncher4rpm-$DownstreamTag.tar.xz" "XIVLauncher4rpm-$DownstreamTag"